Moved bootstrapping to Terraform
For some reason, the bootstrapping, which is just a `kubectl apply`, was done on the
machines with a service that was launched with SSH on every single controller.
This commit aims to change this through the addition of the Kubernetes
provider as well as a GKE autopilot cluster to workaround the
limitations of the Kubernetes provider when it comes to applying
manifests (cf official documentation, application of manifests can only
be done at plan time).